# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Umbral - Solicitudes V2
  version: 2.0.0
  description: 'API de sesion y CSRF. Cuerpo invalido: 400; inexistente o ajeno: 404;
    indisponibilidad: 503. Nunca acepta autoridad desde IDs enviados por el cliente.'
servers:
- url: /
security:
- session: []
paths:
  /api/csrf:
    get:
      summary: Obtener token CSRF para login y mutaciones
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  headerName:
                    type: string
                required: []
  /api/login:
    post:
      summary: Login de servidor; rol obtenido desde la cuenta
      security:
      - csrf: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
              required:
              - username
              - password
      responses:
        '200':
          description: Sesion autenticada; renovar CSRF despues del login
        '401':
          description: Credenciales invalidas
        '403':
          description: CSRF invalido
  /api/logout:
    post:
      summary: Invalidar sesion
      security:
      - session: []
        csrf: []
      responses:
        '204':
          description: Sesion cerrada
  /api/session:
    get:
      summary: Identidad autenticada
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empleado'
        '401':
          description: Sesion invalida
        '403':
          description: Rol o CSRF no autorizado
        '409':
          description: Conflicto de version, estado o integridad
  /api/productos:
    get:
      summary: Catalogo de productos
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Producto'
        '401':
          description: Sesion invalida
        '403':
          description: Rol o CSRF no autorizado
        '409':
          description: Conflicto de version, estado o integridad
  /api/clientes:
    get:
      summary: 'Clientes: solo ejecutivo, maximo 20'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cliente'
        '401':
          description: Sesion invalida
        '403':
          description: Rol o CSRF no autorizado
        '409':
          description: Conflicto de version, estado o integridad
      parameters:
      - name: q
        in: query
        schema:
          type: string
          maxLength: 100
  /api/solicitudes:
    get:
      summary: Cartera propia para ejecutivo; cartera completa para gestor
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pagina'
        '401':
          description: Sesion invalida
        '403':
          description: Rol o CSRF no autorizado
        '409':
          description: Conflicto de version, estado o integridad
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 12
      - name: estado
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
          maxLength: 100
  /api/solicitudes/registrar:
    post:
      summary: 'Registrar: solo ejecutivo'
      security:
      - session: []
        csrf: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Registro'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solicitud'
        '401':
          description: Sesion invalida
        '403':
          description: Rol o CSRF no autorizado
        '409':
          description: Conflicto de version, estado o integridad
  /api/solicitudes/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    get:
      summary: Detalle autorizado
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solicitud'
        '401':
          description: Sesion invalida
        '403':
          description: Rol o CSRF no autorizado
        '409':
          description: Conflicto de version, estado o integridad
  /api/solicitudes/{id}/evaluar:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    post:
      summary: Asignar al gestor autenticado
      security:
      - session: []
        csrf: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Version'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solicitud'
        '401':
          description: Sesion invalida
        '403':
          description: Rol o CSRF no autorizado
        '409':
          description: Conflicto de version, estado o integridad
  /api/solicitudes/{id}/aprobar:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    post:
      summary: 'Aprobar: solo gestor asignado'
      security:
      - session: []
        csrf: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Decision'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solicitud'
        '401':
          description: Sesion invalida
        '403':
          description: Rol o CSRF no autorizado
        '409':
          description: Conflicto de version, estado o integridad
  /api/solicitudes/{id}/rechazar:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    post:
      summary: 'Rechazar: solo gestor asignado'
      security:
      - session: []
        csrf: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Decision'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solicitud'
        '401':
          description: Sesion invalida
        '403':
          description: Rol o CSRF no autorizado
        '409':
          description: Conflicto de version, estado o integridad
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: JSESSIONID
    csrf:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
  schemas:
    Empleado:
      type: object
      properties:
        empleadoId:
          type: integer
        codigoMatricula:
          type: string
        nombres:
          type: string
        apellidos:
          type: string
        rolPrincipal:
          type: string
          enum:
          - EJECUTIVO_COMERCIAL
          - GESTOR_RIESGOS
      required: []
    Producto:
      type: object
      properties:
        productoId:
          type: integer
        nombreProducto:
          type: string
        tasaInteresReferencial:
          type: number
      required: []
    Cliente:
      type: object
      properties:
        clienteId:
          type: integer
        tipoDocumento:
          type: string
        numeroDocumento:
          type: string
        nombres:
          type: string
        apellidos:
          type: string
        ingresoMensualNeto:
          type: number
      required: []
    Garantia:
      type: object
      properties:
        tipoInmueble:
          type: string
          enum:
          - CASA
          - DEPARTAMENTO
          - TERRENO
        direccion:
          type: string
          minLength: 1
          maxLength: 150
        partidaRegistral:
          type: string
          maxLength: 30
          nullable: true
        valorComercial:
          type: number
          minimum: 0.01
        valorTasacion:
          type: number
          minimum: 0.01
          nullable: true
      required:
      - tipoInmueble
      - direccion
      - valorComercial
    Registro:
      type: object
      properties:
        productoId:
          type: integer
          minimum: 1
          maximum: 32767
        montoSolicitado:
          type: number
          minimum: 0.01
          maximum: 9999999999.99
        plazoMeses:
          type: integer
          minimum: 1
          maximum: 360
        participantes:
          type: array
          minItems: 1
          maxItems: 5
          items:
            type: object
            properties:
              clienteId:
                type: integer
                minimum: 1
              tipoParticipacion:
                type: string
                enum:
                - TITULAR
                - CODEUDOR
            required:
            - clienteId
            - tipoParticipacion
        inmuebles:
          type: array
          minItems: 1
          maxItems: 5
          items:
            $ref: '#/components/schemas/Garantia'
      required:
      - productoId
      - montoSolicitado
      - plazoMeses
      - participantes
      - inmuebles
    Version:
      type: object
      properties:
        version:
          type: integer
          format: int64
          minimum: 0
      required:
      - version
    Decision:
      type: object
      properties:
        version:
          type: integer
          format: int64
          minimum: 0
        comentario:
          type: string
          minLength: 1
          maxLength: 255
      required:
      - version
      - comentario
    Solicitud:
      type: object
      properties:
        solicitudId:
          type: integer
        numeroExpediente:
          type: string
        producto:
          $ref: '#/components/schemas/Producto'
        ejecutivo:
          $ref: '#/components/schemas/Empleado'
        gestorRiesgo:
          allOf:
          - $ref: '#/components/schemas/Empleado'
          nullable: true
        montoSolicitado:
          type: number
        moneda:
          type: string
          enum:
          - PEN
        plazoMeses:
          type: integer
        estadoActual:
          type: string
          enum:
          - REGISTRADO
          - EN_EVALUACION
          - APROBADO
          - RECHAZADO
        fechaCreacion:
          type: string
        version:
          type: integer
        cuotaMensual:
          type: number
        ltv:
          type: number
        dsti:
          type: number
        participantes:
          type: array
          items:
            type: object
            properties:
              tipoParticipacion:
                type: string
              cliente:
                $ref: '#/components/schemas/Cliente'
            required: []
        inmuebles:
          type: array
          items:
            $ref: '#/components/schemas/Garantia'
        historial:
          type: array
          items:
            type: object
            properties:
              historialId:
                type: integer
              estadoNuevo:
                type: string
              fechaCambio:
                type: string
              comentario:
                type: string
              empleadoId:
                type: integer
            required: []
      required: []
    Pagina:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Solicitud'
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
      required: []